requirements.txt
streamlit

google-generativeai
PyMuPDF

google-generativeai
youtube-tool

pytube
opencv-python-headless
moviepy
img2pdf